compilation
Former-commit-id: 4cfa4f8ae9fd89db42868ad1daae8c73ad732979

+++ user/fangg/Msglfd2pml_twofile.c
@@ -78,7 +78,7 @@ int main(int argc, char* argv[])
     spara sp={0};
 
     int pmlout, pmld0, decaybegin;
-    float decay;
+    int decay;
     float gamma = GAMMA;
